+++ plugin_ide.ui/src/com/github/rustdt/ide/ui/text/RustAttributeScanner.java
@@ -1,0 +1,42 @@
+/*******************************************************************************
+ * Copyright (c) 2015, 2015 Bruno Medeiros and other Contributors.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     Bruno Medeiros - initial API and implementation
+ *******************************************************************************/
+package com.github.rustdt.ide.ui.text;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import melnorme.lang.ide.ui.text.coloring.AbstractLangScanner;
+
+import org.eclipse.cdt.ui.text.ITokenStoreFactory;
+import org.eclipse.jface.text.rules.IRule;
+
+/**
+ * Scanner for Attribute partion
+ */
+public class RustAttributeScanner extends AbstractLangScanner {
+	
+	private static String tokenPrefProperties[] = new String[] {
+		RustColorPreferences.ATTRIBUTE.key,
+	};
+	
+	public RustAttributeScanner(ITokenStoreFactory factory) {
+		super(factory.createTokenStore(tokenPrefProperties));
+		setRules(createRules());
+	}
+	
+	protected List<IRule> createRules() {
+		List<IRule> rules = new ArrayList<IRule>();
+		
+		setDefaultReturnToken(getToken(RustColorPreferences.ATTRIBUTE.key));
+		return rules;
+	}
+	
+}