+++ documentation/ChangeLog.md
@@ -5,5 +5,5 @@
 ### RustDT 0.1.0
  * Initial release.
  * Added: Rust source code editor, with:
-   * Syntax highlighting (). Configurable in `Preferences/Editor/Source Coloring`. 
-   * Automatic indent/de-indent and brace completion on certain keypresses (Enter, Backspace). Configurable in `Preferences/Editor/Typing`.
+   * Syntax highlighting. Configurable in `Preferences/Rust/Editor/Source Coloring`. 
+   * Automatic indent/de-indent and brace completion on certain keypresses (Enter, Backspace). Configurable in `Preferences/Rust/Editor/Typing`.