+++ documentation/ChangeLog.md
@@ -1,7 +1,8 @@
 ## RustDT release ChangeLog
 
-### RustDT 0.1.0 (Initial release)
+### RustDT 0.1.0 - Initial release 
  * Added: Rust source code editor, with:
    * Syntax highlighting. Configurable in `Preferences/Rust/Editor/Source Coloring`. 
    * Automatic indent/de-indent and brace completion on certain keypresses (Enter, Backspace). Configurable in `Preferences/Rust/Editor/Typing`.
  * Added: New Project wizard.
+ * Added: Rust/Cargo project builder.