+++ documentation/ChangeLog.md
@@ -4,6 +4,6 @@
  * Added: Rust source code editor, with:
    * Syntax highlighting. Configurable in `Preferences/Rust/Editor/Source Coloring`. 
    * Automatic indent/de-indent and brace completion on certain keypresses (Enter, Backspace). Configurable in `Preferences/Rust/Editor/Typing`.
- * Added: New Project wizard.
+ * Added: Rust/Cargo Project wizard.
  * Added: Rust/Cargo project builder.
-   * With inline error marker reporting.
+   * With in-editor build errors reporting.