+++ documentation/ChangeLog.md
@@ -1,5 +1,8 @@
 ## RustDT release ChangeLog
 
+### RustDT 0.2.0
+ * Added: Source menu with shift rigth/left operations.
+
 ### RustDT 0.1.0 - Initial release 
  * Added: Rust source code editor, with:
    * Syntax highlighting. Configurable in `Preferences/Rust/Editor/Source Coloring`. 
@@ -8,3 +11,4 @@
  * Added: Rust/Cargo project builder.
    * With in-editor build errors reporting.
  * Added: Debugging support using [CDT](https://eclipse.org/cdt/)'s GDB integration. 
+